Fixing cast and running s_all

+++ test/utility/misc.c
@@ -616,7 +616,7 @@ testutil_pareto(uint32_t rand, uint64_t range, u_int skew)
     S1 = (-1 / PARETO_SHAPE);
     S2 = range * (skew / 100.0) * (PARETO_SHAPE - 1);
     U = 1 - (double)rand / (double)UINT32_MAX;
-    rand = (uint64_t)((pow(U, S1) - 1) * S2);
+    rand = (pow(U, S1) - 1) * S2;
     /*
      * This Pareto calculation chooses out of range values about
      * 2% of the time, from my testing. That will lead to the
